refactor(logging): merge logger redaction behavior tests into one file
logger-redaction-cache.test.ts and logger-silent-redaction-lazy.test.ts
shared ~25 lines of identical mock/setup boilerplate. Consolidate into
logger-redaction-behavior.test.ts with two describe blocks. No test logic
changes; all 3 assertions are preserved.
